+++ package.json
@@ -18,7 +18,6 @@
     "http-proxy": "~1.5.3",
     "jade": "~1.8.2",
     "kue": "~0.8.11",
-    "kue-ui": "0.0.1",
     "moment": "2.8.4",
     "monk": "0.9.1",
     "passport": "0.2.1",
